got the 64bit version fully built successfully
git-svn-id: https://svn.java.net/svn/com4j~svn/trunk/com4j@385 590aa308-3dd6-4388-9d27-cebafe3e2bdc

+++ native/typelib.cpp
@@ -142,7 +142,7 @@ CTypeDecl::~CTypeDecl() {
 	_ASSERT(m_ThreadID == curThread);
 #endif
 	m_pType->ReleaseTypeAttr(m_pAttr);
-	int r = m_pParent->children.erase(m_pType);
+	size_t r = m_pParent->children.erase(m_pType);
 	_ASSERT(r==1);
 }
